+++ CHANGELOG.md
@@ -3,3 +3,13 @@
 ## 0.0.1
 
 * Initial release.
+
+## 1.0.4
+
+* RookTransmissionConfiguration the url parameter was removed
+  
+* setConfiguration method was added to TransmissionSettings
+
+* enqueueActivityEvent was renamed to fix spelling
+
+* enqueueActivityEvents was renamed to fix spelling